+++ src/main/java/org/osgl/exception/ResourceNotFoundException.java
@@ -1,0 +1,47 @@
+package org.osgl.exception;
+
+/*-
+ * #%L
+ * Java Tool
+ * %%
+ * Copyright (C) 2014 - 2020 OSGL (Open Source General Library)
+ * %%
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * 
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ * 
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * #L%
+ */
+
+/**
+ * A generic exception thrown when a resource tried to access cannot be found.
+ */
+public class ResourceNotFoundException extends UnexpectedException {
+
+    public ResourceNotFoundException() {
+        super("Access Denied");
+    }
+
+    public ResourceNotFoundException(String message) {
+        super(message);
+    }
+
+    public ResourceNotFoundException(String message, Object... args) {
+        super(message, args);
+    }
+
+    public ResourceNotFoundException(Throwable cause) {
+        super(cause);
+    }
+
+    public ResourceNotFoundException(Throwable cause, String message, Object... args) {
+        super(cause, message, args);
+    }
+}